lesson2.1_fix3
